Yield after submitting nc command
Up until now, it was possible that the order of messages submitted by
example scripts to the pipeline was mixed up if all of the `nc` commands
were submitted in one CPU processing slot. This caused the affected
examples to come to wrong results.
To fix this, the processing thread is sent to sleep after each command.
By doing this, the thread yields it processing time, and the `nc`
command gets submitted.
Fixes #256